feat: move auto suggest main component to root formElements
