+++ stories/AutoSuggestInput.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from "@storybook/react";
-import AutoSuggestInput from "~/components/formElements/autoSuggestInput/AutoSuggestInput";
+import AutoSuggestInput from "~/components/formElements/AutoSuggestInput";
 import type { AutoSuggestInputProps } from "~/components/formElements/autoSuggestInput/types";
 import { RVFProvider } from ".storybook/RVFProvider";
 import { reactRouterContext } from ".storybook/reactRouterContext";
